Move bookmarks into Home section picker and add Profile tab
Bookmarks now live between the default sections and user lists in the
Home title menu instead of a dedicated tab. When at least one bookmark
exists, a one-time tip surfaces at the top of the Home view pointing
users at the new entry. The freed tab slot hosts the former More
screen, renamed Profile, so settings are reachable with one tap again.
