+++ SakuraRSS/Views/Home/AllArticlesView.swift
@@ -1,4 +1,5 @@
 import SwiftUI
+import TipKit
 
 enum HomeSection: String, CaseIterable, Identifiable {
     case feed
@@ -37,19 +38,25 @@ enum HomeSection: String, CaseIterable, Identifiable {
 }
 
 /// Represents the selected view in the Home tab title menu.
-/// Can be a static section or a user-created list.
+/// Can be a static section, the bookmarks collection, or a user-created list.
 enum HomeSelection: Hashable, RawRepresentable {
     case section(HomeSection)
+    case bookmarks
     case list(Int64)
 
     var rawValue: String {
         switch self {
         case .section(let section): "section.\(section.rawValue)"
+        case .bookmarks: "bookmarks"
         case .list(let id): "list.\(id)"
         }
     }
 
     init?(rawValue: String) {
+        if rawValue == "bookmarks" {
+            self = .bookmarks
+            return
+        }
         if rawValue.hasPrefix("section.") {
             let sectionRaw = String(rawValue.dropFirst("section.".count))
             if let section = HomeSection(rawValue: sectionRaw) {
@@ -74,6 +81,7 @@ enum HomeSelection: Hashable, RawRepresentable {
     var localizedTitle: String {
         switch self {
         case .section(let section): section.localizedTitle
+        case .bookmarks: String(localized: "Tabs.Bookmarks")
         case .list: ""
         }
     }
@@ -81,6 +89,7 @@ enum HomeSelection: Hashable, RawRepresentable {
     var systemImage: String {
         switch self {
         case .section(let section): section.systemImage
+        case .bookmarks: "bookmark"
         case .list: ""
         }
     }
@@ -126,11 +135,15 @@ struct AllArticlesView: View {
         switch selectedSelection {
         case .section(let section):
             return section.localizedTitle
+        case .bookmarks:
+            return String(localized: "Tabs.Bookmarks")
         case .list(let id):
             return feedManager.lists.first { $0.id == id }?.name
                 ?? String(localized: "Shared.AllArticles")
         }
     }
+
+    private let bookmarksSectionTip = BookmarksSectionTip()
 
     var body: some View {
         Group {
@@ -146,12 +159,21 @@ struct AllArticlesView: View {
                 case .audio:
                     HomeSectionView(section: .audio)
                 }
+            case .bookmarks:
+                BookmarksContentView()
             case .list(let id):
                 if let list = feedManager.lists.first(where: { $0.id == id }) {
                     ListSectionView(list: list)
                 } else {
                     feedTabContent
                 }
+            }
+        }
+        .safeAreaInset(edge: .top, spacing: 0) {
+            if selectedSelection != .bookmarks {
+                TipView(bookmarksSectionTip)
+                    .padding(.horizontal)
+                    .padding(.top, 8)
             }
         }
         .navigationTitle(currentTitle)
@@ -165,6 +187,16 @@ struct AllArticlesView: View {
                 } label: {
                     Label(section.localizedTitle, systemImage: section.systemImage)
                 }
+            }
+
+            Divider()
+            Button {
+                withAnimation(.smooth.speed(2.0)) {
+                    selectedSelection = .bookmarks
+                }
+                bookmarksSectionTip.invalidate(reason: .actionPerformed)
+            } label: {
+                Label("Tabs.Bookmarks", systemImage: "bookmark")
             }
 
             if !feedManager.lists.isEmpty {
@@ -186,6 +218,17 @@ struct AllArticlesView: View {
         .onChange(of: feedManager.lists) {
             validateSelection()
         }
+        .onAppear {
+            refreshBookmarksTip()
+        }
+        .onChange(of: feedManager.dataRevision) {
+            refreshBookmarksTip()
+        }
+    }
+
+    private func refreshBookmarksTip() {
+        let count = (try? DatabaseManager.shared.bookmarkedCount()) ?? 0
+        BookmarksSectionTip.bookmarkCount = count
     }
 
     private func validateSelection() {
@@ -194,6 +237,8 @@ struct AllArticlesView: View {
             if !availableSections.contains(section) {
                 selectedSelection = .section(.feed)
             }
+        case .bookmarks:
+            break
         case .list(let id):
             if !feedManager.lists.contains(where: { $0.id == id }) {
                 selectedSelection = .section(.feed)
